test(gp): split employee component tests into per-component files
Per review feedback: one test file per component (SubmitButton, BackButton,
PersonalDetailsStep, HomeAddressStep, BankAccountStep, FederalTaxesStep,
StateTaxesStep, useEmployeeStepSubmitHandler), with shared mock setup extracted
into tests/helpers.tsx.
